cabal package cleanup
